requirements.txt
fastapi
uvicorn
pillow
python-multipart
numpy
tensorflow==2.19.0